requirements.txt
streamlit>=1.34.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
sentence-transformers>=2.2.2
transformers>=4.40.0
torch>=2.0.0
pypdf>=4.0.0
python-docx>=1.0.0
